+++ customtkinter/windows/widgets/scaling/scaling_base_class.py
@@ -57,13 +57,24 @@ class CTkScalingBaseClass:
     def _get_window_scaling(self) -> float:
         return self.__window_scaling
 
-    def _apply_widget_scaling(self, value: Union[int, float]) -> Union[float]:
+    # Some parts of Tk - notably canvas - are very buggy with floats, because they use locale-dependent parsing
+    # (and thus might not recognize "." as the decimal point)
+    # https://wiki.tcl-lang.org/page/locale
+    # https://github.com/python/cpython/issues/56767
+    # Hence, we must ensure any integer value stays that way
+    def _apply_widget_scaling(self, value: Union[int, float]) -> Union[int, float]:
         assert self.__scaling_type == "widget"
-        return value * self.__widget_scaling
+        if isinstance(value, float):
+            return value * self.__widget_scaling
+        else:
+            return int(value * self.__widget_scaling)
 
-    def _reverse_widget_scaling(self, value: Union[int, float]) -> Union[float]:
+    def _reverse_widget_scaling(self, value: Union[int, float]) -> Union[int, float]:
         assert self.__scaling_type == "widget"
-        return value / self.__widget_scaling
+        if isinstance(value, float):
+            return value / self.__widget_scaling
+        else:
+            return int(value / self.__widget_scaling)
 
     def _apply_window_scaling(self, value: Union[int, float]) -> int:
         assert self.__scaling_type == "window"